requirements/base.txt
Django==4.2.11
django-tenants==3.6.1
psycopg2-binary==2.9.9
redis==5.0.3
celery==5.3.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
python-decouple==3.8
requests==2.31.0
gunicorn==21.2.0
dj-database-url==2.1.0
